refactor(gradle)!: change gradle project struct now trees in lib, save to db and app in app change dependencies add gradle conventions
