+++ app/build.gradle.kts
@@ -1,33 +1,22 @@
 plugins {
-    kotlin("jvm") version "1.8.0"
-    application
+    id("trees.kotlin-application-conventions")
 }
 
 group = "org.example"
 version = "1.0-SNAPSHOT"
 
-repositories {
-    mavenCentral()
-}
-
 val sqliteJdbcVersion: String by project
 
 dependencies {
-    testImplementation("org.jetbrains.kotlin:kotlin-test:1.8.10")
-    testImplementation(platform("org.junit:junit-bom:5.9.2"))
-    testImplementation("org.junit.jupiter:junit-jupiter:5.9.2")
+    implementation(project(":lib"))
     implementation("org.neo4j.driver:neo4j-java-driver:5.7.0")
     implementation("com.fasterxml.jackson.module:jackson-module-kotlin:2.14.2")
     implementation("org.xerial", "sqlite-jdbc", sqliteJdbcVersion)
 }
-
-tasks.test {
-    useJUnitPlatform()
-}
-
-kotlin {
-    jvmToolchain(8)
-}
+//
+//kotlin {
+//    jvmToolchain(8)
+//}
 
 application {
     mainClass.set("MainKt")
