+++ build-logic/src/main/kotlin/trees.kotlin-application-conventions.gradle.kts
@@ -1,0 +1,5 @@
+plugins {
+    id("trees.kotlin-common-conventions")
+
+    application
+}
