+++ build-logic/src/main/kotlin/trees.kotlin-library-conventions.gradle.kts
@@ -1,0 +1,5 @@
+plugins {
+    id("trees.kotlin-common-conventions")
+
+    `java-library`
+}
