+++ lib/src/main/kotlin/AVLTree.kt
@@ -1,5 +1,3 @@
-import abstractTree.BalanceTree
-
 open class AVLTree<Key : Comparable<Key>, Value> : BalanceTree<Key, Value> {
     protected class AVLNode<Key : Comparable<Key>, Value>(
         key: Key,
