+++ lib/src/main/kotlin/Tree.kt
@@ -1,5 +1,3 @@
-package interfaces
-
 interface Tree<Key, Value> {
     fun insert(key: Key, value: Value)
     fun insert(vararg array: Pair<Key, Value>)
